重构了后端状态机逻辑，实现了活动标签和详情，还有签到，踢出活动功能
