+++ backend/Alumni/LogicManager/Constants.py
@@ -1,26 +1,44 @@
 class Constants:
+    '''
+
+
+    '''
+
+    #常量
     UNDEFINED_NUMBER = -1
     MAX_NUMBER = 1145141919810
-    ACTIVITY_STATUS_EXCEPT = 0
-    ACTIVITY_STATUS_BEFORESIGNUP = 1
-    ACTIVITY_STATUS_SIGNUP = 2
-    ACTIVITY_STATUS_SIGNUPPAUSED = 3
-    ACTIVITY_STATUS_SIGNUPSTOPPED = 4
-    ACTIVITY_STATUS_SIGNIN = 5
-    ACTIVITY_STATUS_SIGNINPAUSED = 6
-    ACTIVITY_STATUS_FINISH = 7
+    
+    MIN_MAX_USER = 3
 
+    #活动状态
+    ACTIVITY_STATUS_GLOBAL_EXCEPT = 0
+    ACTIVITY_STATUS_GLOBAL_NORMAL = 1
+    ACTIVITY_STATUS_GLOBAL_FINISH = 2
+    ACTIVITY_STATUS_JOIN_BEFORE = 0
+    ACTIVITY_STATUS_JOIN_CONTINUE = 1
+    ACTIVITY_STATUS_JOIN_PAUSED = 2
+    ACTIVITY_STATUS_JOIN_STOPPED = 3
+    ACTIVITY_STATUS_CHECK_BEFORE = 0
+    ACTIVITY_STATUS_CHECK_CONTINUE = 1
+    ACTIVITY_STATUS_CHECK_PAUSED = 2
+    ACTIVITY_STATUS_CHECK_STOPPED = 3
+
+    #用户状态
+    USER_STATUS_NOT_JOINED = -1
     USER_STATUS_WAITVALIDATE = 0
     USER_STATUS_JOINED = 1
-    USER_STATUS_NOTCHECKED = 2
-    USER_STATUS_CHECKED = 3
-    USER_STATUS_FINISHED = 4
-    USER_STATUS_MISSED = 5
+    USER_STATUS_CHECKED = 2
+    USER_STATUS_FINISHED = 3
+    USER_STATUS_FINISHED_WITHOUT_CHECK = 4
+    USER_STATUS_ABNORMAL = 5
+    USER_STATUS_REFUSED = 6
 
-    USER_ROLE_COMMON = 0
+    #用户权限
+    USER_ROLE_COMMONER = 0
     USER_ROLE_MANAGER = 1
     USER_ROLE_CREATOR = 2
 
+    #错误码
     ERROR_CODE_NOT_STORED = 101
     ERROR_CODE_RECOMMEND = 201
     ERROR_CODE_NO_REASON = 211
@@ -34,6 +52,7 @@ class Constants:
     ERROR_CODE_NETWORK_ERROR = 506
     ERROR_CODE_DENIAL = 507
 
+    #规则类型
     GLOBAL_RULE_ACCEPT = 0
     GLOBAL_RULE_AUDIT = 1
     GLOBAL_RULE_REJECT = 2
@@ -42,5 +61,6 @@ class Constants:
     ADVANCED_RULE_AUDIT = 1
     ADVANCED_RULE_REJECT = 2
 
+    #搜索相关
     MAX_SEARCH_RESULT = 10
     MAX_RECOMMEND_WORD = 5
